Fix k8s and flannel version placeholders missing
__KUBERNETES_VERSION__ and __FLANNEL_VERSION__ placeholders are
populated in defaults.env when building the package. Since these files
are embedded in nodelet before the replacement occurs these placeholders
were left as-is. Creating a cluster after the directory is extracted using
`/opt/pf9/nodelet/nodeletd init` breaks because of such placeholders.
This commit addresses this issue by always populating the placeholders.